Moved steps methods to separate class
